fix: google analytics integration
