+++ pages/_app.tsx
@@ -1,9 +1,11 @@
-import '../src/styles/globals.css';
+import { useEffect } from 'react';
 import '@fortawesome/fontawesome-free/css/all.css';
 import type { AppProps } from 'next/app';
 import AOS from 'aos';
 import 'aos/dist/aos.css';
-import { useEffect } from 'react'; // You can also use <link> for styles
+
+import '../src/styles/globals.css';
+
 function MyApp({ Component, pageProps }: AppProps) {
   useEffect(() => {
     AOS.init({
@@ -12,7 +14,12 @@ function MyApp({ Component, pageProps }: AppProps) {
     });
   }, []);
 
-  return <Component {...pageProps} />;
+  return (
+    <>
+      <title>Hydranet: Bringing Bitcoin to DeFi</title>
+      <Component {...pageProps} />
+    </>
+  );
 }
 
 export default MyApp;
