+++ pages/index.tsx
@@ -1,5 +1,4 @@
 import type { NextPage } from 'next';
-import Head from 'next/head';
 
 import Layout from '../src/components/Templates/Layout';
 import ProductsSection from '../src/components/Templates/ProductsSection';
@@ -7,17 +6,9 @@ import RoadmapSection from '../src/components/Templates/RoadmapSection';
 import ArticlesSection from '../src/components/Templates/ArticlesSection';
 import AboutSection from '../src/components/Templates/AboutSection';
 import Landing from '../src/components/Templates/Landing';
-import MetaTags from '../src/components/Atoms/MetaTags';
-import GoogleAnalytics from '../src/components/Atoms/GoogleAnalytics';
 
 const Home: NextPage = () => (
   <>
-    <Head>
-      <title>Hydranet: Bringing Bitcoin to DeFi</title>
-      <MetaTags />
-      <link rel="icon" href="/favicon.ico" />
-      <GoogleAnalytics />
-    </Head>
     <Layout>
       <Landing />
       <AboutSection id={'about'} />
